# backend/contracts/openapi/v1/provider.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAKEIN Provider API",
    "version": "1.0.0",
    "description": "Existing provider-owner and provider-branch API contract. Every operation requires JWT authentication plus the server-side provider account, branch, and menu permission checks identified by x-provider-permission."
  },
  "servers": [
    {
      "url": "https://api.takein.id/api",
      "description": "Production API ingress"
    }
  ],
  "paths": {
...
    "/provider/services": {
      "get": {
        "operationId": "listProviderService",
        "summary": "List provider services",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "services",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                10,
                25,
                50,
                100
              ]
            }
          }
        ]
      },
      "post": {
        "operationId": "createProviderService",
        "summary": "Create provider service",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "services",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ServiceWriteRequest"
              }
            }
          }
        }
      }
    },
...
    "/provider/services/{service}/branch": {
      "put": {
        "operationId": "updateProviderServiceBranches",
        "summary": "Update service branch assignments",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "services",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceBranchUpdateRequest"
              }
            }
          }
        }
      }
    },
    "/provider/services/{service}/gallery": {
      "put": {
        "operationId": "updateProviderServiceGallery",
        "summary": "Update service gallery media",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "services",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ServiceGalleryUpdateRequest"
              }
            }
          }
        }
      }
    },
    "/provider/services/{service}/toggle-status": {
      "patch": {
        "operationId": "toggleProviderServiceStatus",
        "summary": "Toggle service active status",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "services",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
...
      "ServiceBranchUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "branch_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer"
            }
          }
        }
      },
      "ServiceGalleryUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "gallery_image": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary",
            "description": "Allowed media: JPEG, PNG, or WebP; maximum 2 MiB."
          },
          "video_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "maxLength": 255
          }
        }
      },
...
    }
  }
}